Added jsdoc and removed some files no longer used
Added jsdoc and removed some files no longer used
