+++ package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "start": "serverless offline start --stage='localdev' --skipCacheInvalidation",
     "test": "jest",
-    "test:api:unit": "NODE_ENV=test ./node_modules/mocha/bin/mocha -r dotenv/config ./tests/api/unit/**/*.test.mjs",
+    "test:api:unit": "NODE_ENV=test jest ./__tests__/api/unit/**/*.spec.js",
     "lint": "node node_modules/eslint/bin/eslint.js .",
     "fixtures": "fixtures ./fixtures --config ./ormconfig.js --sync --require=ts-node/register --require=tsconfig-paths/register",
     "typeorm": "ts-node ./node_modules/typeorm/cli.js",
@@ -25,14 +25,12 @@
     "babel-jest": "24.9.0",
     "babel-loader": "8.0.6",
     "babel-plugin-source-map-support": "2.1.1",
-    "chai": "^4.3.4",
     "eslint": "6.8.0",
     "eslint-config-airbnb": "18.0.1",
     "eslint-plugin-import": "2.20.0",
     "eslint-plugin-jsx-a11y": "6.2.3",
     "eslint-plugin-react": "7.18.0",
     "jest": "24.9.0",
-    "mocha": "^9.1.3",
     "regenerator-runtime": "0.13.3",
     "serverless": "1.61.1",
     "serverless-dotenv-plugin": "^3.2.0",
