+++ src/di-container.js
@@ -15,7 +15,7 @@ import S3Service from './services/s3.service'
 import { Headmate } from './entities/Headmate'
 
 export const logger = Logger({})
-export const s3Service = S3Service({ config, s3Client, PutObjectCommand, DeleteObjectCommand, getSignedUrl })
+export const s3Service = S3Service({ s3Client, PutObjectCommand, DeleteObjectCommand, getSignedUrl })
 export const userService = UserService({  })
 export const headmateService = HeadmateService({ config, entityValidate, s3Service, getConnection, closeConnection, Headmate })
 export const authController = AuthController({ config, jwt, logger, userService })
