--- tests/fixtures.mjs
@@ -1,3 +1,0 @@
-export async function mochaGlobalSetup() {
-
-}
